Make createJob a public method
This function is beneficial if you want to instantiate a job with
the job plugin manager and you have the queue instance present. The
other method is to pull the job plugin manager manually and set the
content and metadata yourself. This can be repeated for every job
you need to instantiate. So it's better to get this feature public.

The job instances must be created because only the instance (and no
string name) can be pushed into the queue. A second option could be
to allow jobs to be pushed with only string names:

`public function pushByName($name, $content, $metadata, $options)`

You can check existence of the job name within the job plugin manager:

`$this->jobPluginManager->has($name)`