More job serialization out of the job class
The job should not be able to serialize itself why it doesn't
contain the logic to deserialize (ie instantiate) iself from
this string. Otherwise, behaviour and control are mixed between
different layers.

Now, the (de)serialization is done in the queue. This also eases
the usage of the queues as they do not have to repeat the
extraction code. For special queue usage, it is possible to add
additional metadata during creation of the job.