Deprecate Interop\Container\ContainerInterface
Starting Feb. 13th 2017, container-interop is officially deprecated in
favor of PSR-11.

Given that the Interop's readme starts off with the statement above,
then it makes sense to deprecate the package in favour of
Psr\Container\ContainerInterface, which this change implements.

I was motivated to make the change as I'm trying to use the package with
a small Slim Framework project I'm working on, and was restricted from
doing so, because of the project using Interop\Container\ContainerInterface.
After doing a bit of hunting, I came across the above statement in the
README, so decided to create this change.